Delegating merging defaults to createMaterialPalette, updating tests
