+++ src/__tests__/createMaterialPalette.test.js
@@ -14,7 +14,11 @@ import resolveAssetSource
   from 'react-native/Libraries/Image/resolveAssetSource';
 import createPalette from '../createMaterialPalette';
 import validateCreatePalette from '../utils/validateCreatePalette';
-import { defaultLightSwatch, defaultOptions } from '../constants/defaults';
+import {
+  defaultLightSwatch,
+  defaultOptions,
+  defaultSwatches,
+} from '../constants/defaults';
 
 describe('createMaterialPalette', () => {
   beforeEach(() => {
@@ -67,12 +71,122 @@ describe('createMaterialPalette', () => {
         region: defaultOptions.region,
         maximumColorCount: defaultOptions.maximumColorCount,
       });
-      expect(palette.lightVibrant).toBeNull();
+      expect(palette.lightVibrant).toEqual(defaultSwatches.lightVibrant);
       expect(palette.darkMuted).toEqual({
         color: 'green',
         population: 20,
         bodyTextColor: 'red',
         titleTextColor: 'red',
+      });
+    });
+  });
+
+  describe('Merge with defaults', () => {
+    it('should merge palette with globals when custom defaults are not provided, for the types specified', async () => {
+      NativeModules.MaterialPalette.createMaterialPalette.mockImplementation(
+        () =>
+          Promise.resolve({
+            vibrant: {
+              color: 'green',
+              bodyTextColor: 'red',
+              titleTextColor: 'red',
+              population: 20,
+            },
+            muted: null,
+          }),
+      );
+
+      expect(
+        await createPalette(0, { types: ['vibrant', 'muted'] }, undefined),
+      ).toEqual({
+        vibrant: {
+          color: 'green',
+          bodyTextColor: 'red',
+          titleTextColor: 'red',
+          population: 20,
+        },
+        muted: defaultSwatches.muted,
+      });
+    });
+
+    it('should merge palette with globals when custom defaults contain a wrong profile, for the types specified', async () => {
+      NativeModules.MaterialPalette.createMaterialPalette.mockImplementation(
+        () =>
+          Promise.resolve({
+            vibrant: defaultSwatches.vibrant,
+          }),
+      );
+
+      expect(
+        await createPalette(
+          0,
+          { type: ['vibrant', 'muted'] },
+          { darkMuted: null },
+        ),
+      ).toEqual({
+        vibrant: defaultSwatches.vibrant,
+        darkMuted: defaultSwatches.darkMuted,
+      });
+    });
+
+    it('should merge palette with both globals and custom defaults, for the types specified', async () => {
+      NativeModules.MaterialPalette.createMaterialPalette.mockImplementation(
+        () =>
+          Promise.resolve({
+            muted: {
+              color: 'green',
+              bodyTextColor: 'red',
+              titleTextColor: 'red',
+              population: 20,
+            },
+            darkMuted: {
+              color: 'yellow',
+              bodyTextColor: 'blue',
+              titleTextColor: 'blue',
+              population: 40,
+            },
+            lightVibrant: null,
+            darkVibrant: null,
+          }),
+      );
+
+      expect(
+        await createPalette(
+          0,
+          { type: ['muted', 'darkMuted', 'lightVibrant', 'darkVibrant'] },
+          {
+            darkMuted: {
+              color: 'orange',
+              bodyTextColor: 'purple',
+              titleTextColor: 'purple',
+            },
+            lightVibrant: {
+              color: 'orange',
+              bodyTextColor: 'purple',
+              titleTextColor: 'purple',
+            },
+          },
+        ),
+      ).toEqual({
+        muted: {
+          color: 'green',
+          bodyTextColor: 'red',
+          titleTextColor: 'red',
+          population: 20,
+        },
+        darkMuted: {
+          color: 'yellow',
+          bodyTextColor: 'blue',
+          titleTextColor: 'blue',
+          population: 40,
+        },
+        lightVibrant: {
+          color: 'orange',
+          bodyTextColor: 'purple',
+          titleTextColor: 'purple',
+          population: 0,
+        },
+        darkVibrant: defaultSwatches.darkVibrant,
       });
     });
   });
