+++ src/createMaterialPalette.js
@@ -4,13 +4,67 @@ import { NativeModules } from 'react-native';
 import resolveAssetSource
   from 'react-native/Libraries/Image/resolveAssetSource';
 import isEqual from 'lodash/isEqual';
-import { defaultOptions, defaultLightSwatch } from './constants/defaults';
+import {
+  defaultOptions,
+  defaultLightSwatch,
+  defaultSwatches,
+} from './constants/defaults';
 import validateCreatePalette from './utils/validateCreatePalette';
-import type { Image, PaletteInstance, Options, ColorProfile } from './types';
+import type {
+  Image,
+  PaletteInstance,
+  Options,
+  ColorProfile,
+  PaletteDefaults,
+} from './types';
+
+function mergeWithDefaults(
+  palette: PaletteInstance,
+  customDefaults: PaletteDefaults = {},
+) {
+  const globalDefaultsForTypesProvided = ((Object.keys(
+    palette,
+  ): any): ColorProfile[]).reduce(
+    (acc, profile) => ({
+      ...acc,
+      [profile]: defaultSwatches[profile],
+    }),
+    {},
+  );
+
+  const defaults = {
+    ...globalDefaultsForTypesProvided,
+    ...((Object.keys(customDefaults): any): ColorProfile[]).reduce(
+      (acc: *, profile: ColorProfile) => ({
+        ...acc,
+        [profile]: {
+          ...(customDefaults && customDefaults[profile]
+            ? customDefaults[profile]
+            : defaultSwatches[profile]),
+          population: 0,
+        },
+      }),
+      {},
+    ),
+  };
+  return {
+    ...defaults,
+    ...((Object.keys(palette): any): ColorProfile[])
+      .filter((profile: ColorProfile) => !!palette[profile]) // Stripping out unavailable profiles
+      .reduce(
+        (acc: *, profile: ColorProfile) => ({
+          ...acc,
+          [profile]: palette[profile],
+        }),
+        {},
+      ),
+  };
+}
 
 export default (async function createMaterialPalette(
   image: Image,
   options?: Options = {},
+  defaults: PaletteDefaults,
 ): Promise<PaletteInstance> {
   validateCreatePalette(image, options);
   const {
@@ -34,5 +88,5 @@ export default (async function createMaterialPalette(
       paletteInstance[profile] = null;
     }
   });
-  return paletteInstance;
+  return mergeWithDefaults(paletteInstance, defaults);
 });
